Clean up project to reduce warnings

+++ VoodooI2C/VoodooI2C.cpp
@@ -593,7 +593,7 @@ int VoodooI2C::xferI2C(I2CBus* _dev, i2c_msg *msgs, int num) {
     
     nanoseconds_to_absolutetime(10000, &abstime);
     
-    sleep = _dev->commandGate->commandSleep(&_dev->commandComplete, abstime);
+    sleep = _dev->commandGate->commandSleep(&_dev->commandComplete, (UInt32)abstime);
     
                                             
     if ( sleep == THREAD_TIMED_OUT ) {
